Deleting an unexisting layergroup throws a 500
We can safely ignore 404 errors when deleting, but
a 500 we can actually not automatically ignore. 
For some reason the layergroup returns an incorrect
error-code when deleting a non-existing layergroup. 

Make the code a little more robust and behave as
expected by first doing a get and skipping the delete
if the layer does not exist before.